move sbom telemetry to sbom package
