+++ pkg/sbom/telemetry/telemetry.go
@@ -13,6 +13,8 @@ const (
 	subsystem = "sbom"
 )
 
+var commonOpts = telemetry.Options{NoDoubleUnderscoreSep: true}
+
 var (
 	// SBOMAttempts tracks sbom collection attempts.
 	SBOMAttempts = telemetry.NewCounterWithOpts(
@@ -20,7 +22,7 @@ var (
 		"sbom_attempts",
 		[]string{"source", "type"},
 		"Number of sbom failures by (source, type)",
-		telemetry.Options{NoDoubleUnderscoreSep: true},
+		commonOpts,
 	)
 	// SBOMFailures tracks sbom collection attempts that fail.
 	SBOMFailures = telemetry.NewCounterWithOpts(
@@ -28,7 +30,7 @@ var (
 		"sbom_errors",
 		[]string{"source", "type", "reason"},
 		"Number of sbom failures by (source, type, reason)",
-		telemetry.Options{NoDoubleUnderscoreSep: true},
+		commonOpts,
 	)
 
 	// SBOMGenerationDuration measures the time that it takes to generate SBOMs
@@ -39,7 +41,7 @@ var (
 		[]string{},
 		"SBOM generation duration (in seconds)",
 		[]float64{10, 30, 60, 120, 180, 240, 300, 360, 420, 480, 540, 600},
-		telemetry.Options{NoDoubleUnderscoreSep: true},
+		commonOpts,
 	)
 
 	// SBOMCacheMemSize size in memory of the cache used for SBOM collection
@@ -48,15 +50,60 @@ var (
 		"sbom_cache_mem_size",
 		[]string{},
 		"SBOM cache size in memory (in bytes)",
-		telemetry.Options{NoDoubleUnderscoreSep: true},
+		commonOpts,
 	)
 
-	// SBOMCacheDiskSize size in disk of the cache used for SBOM collection
+	// SBOMCacheDiskSize size in disk of the custom cache used for SBOM collection
 	SBOMCacheDiskSize = telemetry.NewGaugeWithOpts(
 		subsystem,
 		"sbom_cache_disk_size",
 		[]string{},
-		"SBOM cache size in disk (in bytes)",
-		telemetry.Options{NoDoubleUnderscoreSep: true},
+		"SBOM size in disk of the custom cache (in bytes)",
+		commonOpts,
+	)
+
+	// SBOM number of cache keys stored in memory
+	SBOMCacheEntries = telemetry.NewGaugeWithOpts(
+		subsystem,
+		"sbom_cached_keys",
+		[]string{},
+		"Number of cache keys stored in memory",
+		commonOpts,
+	)
+
+	// SBOMCachedObjectSize total size of cached objects in disk (in bytes) used for SBOM collection
+	SBOMCachedObjectSize = telemetry.NewGaugeWithOpts(
+		subsystem,
+		"sbom_cached_object_size",
+		[]string{},
+		"SBOM total size of cached objects in disk (in bytes)",
+		commonOpts,
+	)
+
+	// SBOMCacheHits number of cache hits during SBOM collection
+	SBOMCacheHits = telemetry.NewCounterWithOpts(
+		subsystem,
+		"sbom_cache_hits_total",
+		[]string{},
+		"SBOM total number of cache hits during SBOM collection",
+		commonOpts,
+	)
+
+	// SBOMCacheMisses number of cache misses during SBOM collection
+	SBOMCacheMisses = telemetry.NewCounterWithOpts(
+		subsystem,
+		"sbom_cache_misses_total",
+		[]string{},
+		"SBOM total number of cache misses during SBOM collection",
+		commonOpts,
+	)
+
+	// SBOMCacheEvicts number of cache evicts during SBOM collection
+	SBOMCacheEvicts = telemetry.NewCounterWithOpts(
+		subsystem,
+		"sbom_cache_evicts_total",
+		[]string{},
+		"SBOM total number of cache misses during SBOM collection",
+		commonOpts,
 	)
 )
