+++ pkg/workloadmeta/telemetry/telemetry.go
@@ -107,58 +107,4 @@ var (
 		[]float64{10, 30, 60, 120, 180, 240, 300, 360, 420, 480, 540, 600},
 		commonOpts,
 	)
-
-	// SBOM number of cache keys stored in memory
-	SBOMCacheEntries = telemetry.NewGaugeWithOpts(
-		subsystem,
-		"sbom_cached_keys",
-		[]string{},
-		"Number of cache keys stored in memory",
-		commonOpts,
-	)
-
-	// SBOMCacheDiskSize size in disk of the custom cache used for SBOM collection
-	SBOMCacheDiskSize = telemetry.NewGaugeWithOpts(
-		subsystem,
-		"sbom_cache_disk_size",
-		[]string{},
-		"SBOM size in disk of the custom cache (in bytes)",
-		commonOpts,
-	)
-
-	// SBOMCachedObjectSize total size of cached objects in disk (in bytes) used for SBOM collection
-	SBOMCachedObjectSize = telemetry.NewGaugeWithOpts(
-		subsystem,
-		"sbom_cached_object_size",
-		[]string{},
-		"SBOM total size of cached objects in disk (in bytes)",
-		commonOpts,
-	)
-
-	// SBOMCacheHits number of cache hits during SBOM collection
-	SBOMCacheHits = telemetry.NewCounterWithOpts(
-		subsystem,
-		"sbom_cache_hits_total",
-		[]string{},
-		"SBOM total number of cache hits during SBOM collection",
-		commonOpts,
-	)
-
-	// SBOMCacheMisses number of cache misses during SBOM collection
-	SBOMCacheMisses = telemetry.NewCounterWithOpts(
-		subsystem,
-		"sbom_cache_misses_total",
-		[]string{},
-		"SBOM total number of cache misses during SBOM collection",
-		commonOpts,
-	)
-
-	// SBOMCacheEvicts number of cache evicts during SBOM collection
-	SBOMCacheEvicts = telemetry.NewCounterWithOpts(
-		subsystem,
-		"sbom_cache_evicts_total",
-		[]string{},
-		"SBOM total number of cache misses during SBOM collection",
-		commonOpts,
-	)
 )
